#16, need to migrate to long lng completely in vm_list.cpp:33

+++ src/common/common.h
@@ -5,13 +5,13 @@
 #include <string>
 #include <vector>
 
-static constexpr double BASE = 1000;
+static constexpr long long BASE = 1000;
 static const std::vector<std::string> SizeSuffixes = {"B", "KB", "MB", "GB", "TB"};
 static const std::vector<std::string> PerformanceSuffixes = {"F", "KF", "MF", "GF", "TF"};
 
 static constexpr int DefaultCores = 1;
 static constexpr int DefaultMemory = 0;
 
-double ParseSize(const std::string& size, const std::vector<std::string>& suffixes);
+long long ParseSize(const std::string& size, const std::vector<std::string>& suffixes);
 
 void TransformHostsProps();
